+++ src/dcos_e2e_cli/dcos_aws/commands/sync.py
@@ -6,6 +6,7 @@ from pathlib import Path
 
 import click
 
+from dcos_e2e_cli._vendor.dcos_installer_tools import DCOSVariant
 from dcos_e2e_cli.common.arguments import dcos_checkout_dir_argument
 from dcos_e2e_cli.common.options import (
     existing_cluster_id_option,
@@ -52,8 +53,13 @@ def sync_code(
         cluster_id=cluster_id,
         aws_region=aws_region,
     )
+    dcos_variant = {
+        True: DCOSVariant.ENTERPRISE,
+        False: DCOSVariant.OSS,
+    }[cluster_instances.is_enterprise]
     cluster = cluster_instances.cluster
     sync_code_to_masters(
         cluster=cluster,
         dcos_checkout_dir=Path(dcos_checkout_dir),
+        dcos_variant=dcos_variant,
     )