+++ src/dcos_e2e_cli/dcos_vagrant/commands/run_command.py
@@ -8,6 +8,7 @@ from typing import Dict, Optional, Tuple
 import click
 
 from dcos_e2e.node import Node, Transport
+from dcos_e2e_cli._vendor.dcos_installer_tools import DCOSVariant
 from dcos_e2e_cli.common.arguments import node_args_argument
 from dcos_e2e_cli.common.options import (
     dcos_login_pw_option,
@@ -122,11 +123,16 @@ def run(
     host = _get_node(cluster_id=cluster_id, node_reference=node)
     cluster_vms = ClusterVMs(cluster_id=cluster_id)
     cluster = cluster_vms.cluster
+    dcos_variant = {
+        True: DCOSVariant.ENTERPRISE,
+        False: DCOSVariant.OSS,
+    }[cluster_vms.is_enterprise]
 
     if sync_dir is not None:
         sync_code_to_masters(
             cluster=cluster,
             dcos_checkout_dir=sync_dir,
+            dcos_variant=dcos_variant,
         )
 
     run_command(